moved invest log markup function to a module. #212.
